+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed an issue returning internal server error when operational reports are generated
 - Fixed an issue related to rewriting a k8s platform during creation if the type is not SELF_MANAGED
 - Fixed an issue with resource collection when pod killed during memory limit exceeded
+- Fixed license expiration updating in case of 404 status code returned from the license sync request
 
 ## [5.15.0] - 2025-11-06
 - Added hiding expired resource exceptions